hub-root/backend/file_convert_service/app/file_convert/serializers.py
from rest_framework import serializers
from .models import FileConvert


class FileConvertSerializer(serializers.ModelSerializer):
    class Meta:
        model = FileConvert
        fields = '__all__'