hub-root/backend/file_convert_service/app/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.33.10
botocore==1.33.10
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.4
Markdown==3.5.1
packaging==23.2
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
s3transfer==0.8.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.12
zope.interface==6.1
